fix: removendo interface com duplicidade, e corrigindo os metodos que implementavam a interface deletada

+++ backend/src/com/dbc/repository/ProfissionalMentorRepository.java
@@ -18,7 +18,7 @@ public class ProfissionalMentorRepository implements IRepository<Long, Profissio
         ResultSet res = stmt.executeQuery(sql);
 
         if (res.next()) {
-            return Long.valueOf(res.getInt("mysequence"));
+            return res.getLong("mysequence");
         }
 
         return null;
